Move loading indicator view to shared scope
