+++ Tireless/Shared/Loading View/IndicatorView.swift
@@ -1,0 +1,51 @@
+//
+//  IndicatorView.swift
+//  Tireless
+//
+//  Created by Liam on 2024/1/5.
+//
+
+import SwiftUI
+
+struct IndicatorView: View {
+
+    let count: Int
+
+    var body: some View {
+        GeometryReader { geometry in
+            ForEach(0..<count, id: \.self) { index in
+                IndicatorItemView(index: index, size: geometry.size)
+            }
+            .frame(width: geometry.size.width, height: geometry.size.height)
+        }
+    }
+}
+
+struct IndicatorItemView: View {
+
+    let index: Int
+    let size: CGSize
+
+    @State private var scale: CGFloat = 0
+    @State private var rotation: Double = 0
+
+    var body: some View {
+        let animation = Animation
+            .timingCurve(0.5, 0.15 + Double(index) / 5, 0.25, 1, duration: 1.5)
+            .repeatForever(autoreverses: false)
+
+        return Circle()
+            .frame(width: size.width / 5, height: size.height / 5)
+            .scaleEffect(scale)
+            .offset(y: size.width / 10 - size.height / 2)
+            .rotationEffect(.degrees(rotation))
+            .onAppear {
+                rotation = 0
+                scale = (5 - CGFloat(index)) / 5
+                withAnimation(animation) {
+                    rotation = 360
+                    scale = (1 + CGFloat(index)) / 5
+                }
+            }
+    }
+}
